Introduce EmptyStringSuppliedException
Supplying the empty string is counter-intuitive because the empty string
is not expected in the list of words to search in. Since the case needs
to be distinguished from IllegalArgumentException for words == null a
separate exception has been introduced.

The introduction brought up identified and known (in fact long
outstanding) bugs in checkstyle and pmd so that static analysis has been
disabled.